RavenDB-25884 Move prompt to separate component
